Implementando o controle e gerenciamento de sessão e usuários (adm e comum)

+++ CidadeUnida/Controllers/ContatoController.cs
@@ -1,6 +1,8 @@
 ﻿using CidadeUnida.Models;
 using CidadeUnida.Repositories.ADO.SQL_Server;
+using CidadeUnida.ViewModels;
 using Microsoft.AspNetCore.Mvc;
+using Newtonsoft.Json;
 
 namespace CidadeUnida.Controllers
 {
@@ -12,30 +14,50 @@ namespace CidadeUnida.Controllers
         {
             repository = new ContatoDAO(
                                     configuration.GetConnectionString(
-                                       Configurations.AppSettings.GetKeyConnectionString())); 
+                                       Configurations.AppSettings.GetKeyConnectionString()));
         }
 
         // GET: Contato
         public IActionResult Index()
         {
+            var permissao = VerificarPermissao();
+            if (permissao != null)
+            {
+                return permissao; // Redireciona se o usuário não tiver permissão
+            }
+
             List<Contato> contatos = repository.GetAll();
             return View(contatos);
+
         }
 
         // GET: Contato/Details/5
         public IActionResult Details(int id)
         {
+            var permissao = VerificarPermissao();
+            if (permissao != null)
+            {
+                return permissao; // Redireciona se o usuário não tiver permissão
+            }
+
             Contato contato = repository.GetByIdContato(id);
             if (contato == null)
             {
                 return NotFound();
             }
             return View(contato);
+
         }
 
         // GET: Contato/Create
         public IActionResult Create()
         {
+            var permissao = VerificarPermissao();
+            if (permissao != null)
+            {
+                return permissao; // Redireciona se o usuário não tiver permissão
+            }
+
             return View();
         }
 
@@ -43,6 +65,12 @@ namespace CidadeUnida.Controllers
         [HttpPost]
         public IActionResult Create(Contato contato)
         {
+            var permissao = VerificarPermissao();
+            if (permissao != null)
+            {
+                return permissao; // Redireciona se o usuário não tiver permissão
+            }
+
             try
             {
                 repository.Add(contato);
@@ -57,6 +85,12 @@ namespace CidadeUnida.Controllers
         // GET: Contato/Edit/5
         public IActionResult Edit(int id)
         {
+            var permissao = VerificarPermissao();
+            if (permissao != null)
+            {
+                return permissao; // Redireciona se o usuário não tiver permissão
+            }
+
             Contato contato = repository.GetByIdContato(id);
             if (contato == null)
             {
@@ -70,6 +104,12 @@ namespace CidadeUnida.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Edit(int id, Contato contato)
         {
+            var permissao = VerificarPermissao();
+            if (permissao != null)
+            {
+                return permissao; // Redireciona se o usuário não tiver permissão
+            }
+
             try
             {
                 repository.Update(id, contato);
@@ -85,8 +125,33 @@ namespace CidadeUnida.Controllers
         [HttpGet]
         public IActionResult Delete(int id)
         {
+            var permissao = VerificarPermissao();
+            if (permissao != null)
+            {
+                return permissao; // Redireciona se o usuário não tiver permissão
+            }
+
             repository.Delete(id);
             return RedirectToAction(nameof(Index));
+        }
+
+        private IActionResult VerificarPermissao()
+        {
+            var loginTokenJson = HttpContext.Session.GetString("login");
+            if (loginTokenJson == null)
+            {
+                TempData["ErrorMessage"] = "É necessário estar logado para realizar essa operação!";
+
+                return RedirectToAction("Login", "Login");
+            }
+
+            var login = JsonConvert.DeserializeObject<LoginViewModel>(loginTokenJson);
+            if (!login.IsAdm)
+            {
+                return RedirectToAction("AcessoNegado", "Home");
+            }
+
+            return null; // Indica que o acesso está permitido
         }
     }
 }
